refactor: Move Windows binary loader config classes into separate files
