+++ src/CoreHook.BinaryInjection/Loader/BinaryLoaderConfig.Windows.cs
@@ -2,20 +2,6 @@
 
 namespace CoreHook.BinaryInjection.Loader
 {
-    public sealed partial class BinaryLoaderHostConfig
-    {
-        public static string CoreCLRStartFunction
-            => "StartCoreCLR";
-        public static string CoreCLRExecuteManagedFunction
-            => "ExecuteAssemblyFunction";
-    }
-
-    public interface IBinaryLoaderConfig
-    {
-        int MaxPathLength { get; }
-        Encoding PathEncoding { get; }
-    }
-
     public sealed partial class BinaryLoaderConfig : IBinaryLoaderConfig
     {
         public int MaxPathLength => 260;
