+++ src/CoreHook.BinaryInjection/Loader/BinaryLoaderHostConfig.cs
@@ -1,0 +1,11 @@
+﻿
+namespace CoreHook.BinaryInjection.Loader
+{
+    public sealed partial class BinaryLoaderHostConfig
+    {
+        public static string CoreCLRStartFunction
+            => "StartCoreCLR";
+        public static string CoreCLRExecuteManagedFunction
+            => "ExecuteAssemblyFunction";
+    }
+}
