+++ src/CoreHook.BinaryInjection/Loader/IBinaryLoaderConfig.cs
@@ -1,0 +1,10 @@
+﻿using System.Text;
+
+namespace CoreHook.BinaryInjection.Loader
+{
+    public interface IBinaryLoaderConfig
+    {
+        int MaxPathLength { get; }
+        Encoding PathEncoding { get; }
+    }
+}
